+++ VS.Squirrel/SquirrelPackagerPackage.cs
@@ -41,8 +41,9 @@ namespace AutoSquirrel
     [InstalledProductRegistration("#110", "#112", "1.0", IconResourceID = 400)] // Info on this package for Help/About
     [ProvideMenuResource("Menus.ctmenu", 1)]
     [ProvideToolWindow(typeof(SquirrelPackager))]
-    [ProvideToolWindowVisibility(typeof(SquirrelPackager), VSConstants.UICONTEXT.SolutionExists_string)]
+    [ProvideToolWindowVisibility(typeof(SquirrelPackager), VSConstants.UICONTEXT.SolutionExistsAndFullyLoaded_string)]
     [Guid(SquirrelPackagerPackage.PackageGuidString)]
+    [ProvideAutoLoad(UIContextGuids80.SolutionExists)]
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1650:ElementDocumentationMustBeSpelledCorrectly", Justification = "pkgdef, VS and vsixmanifest are valid VS terms")]
     public sealed class SquirrelPackagerPackage : Package
     {